+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>org.javalaboratories</groupId>
     <artifactId>java-extensions</artifactId>
-    <version>2.0.0.0-RELEASE</version>
+    <version>2.0.1.0-RELEASE</version>
 
     <name>${project.groupId}:${project.artifactId}</name>
     <description>This is a library of utilities to encourage functional
