+++ src/main/java/org/javalaboratories/core/cryptography/RsaMessageVerifier.java
@@ -1,7 +1,7 @@
 package org.javalaboratories.core.cryptography;
 
-import org.javalaboratories.core.cryptography.transport.JsonHelper;
-import org.javalaboratories.core.cryptography.transport.Message;
+import org.javalaboratories.core.cryptography.json.JsonHelper;
+import org.javalaboratories.core.cryptography.json.Message;
 
 import java.io.File;
 import java.security.PrivateKey;
