+++ src/main/java/org/javalaboratories/core/cryptography/json/JsonHelper.java
@@ -13,7 +13,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package org.javalaboratories.core.cryptography.transport;
+package org.javalaboratories.core.cryptography.json;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
