+++ src/main/java/org/javalaboratories/core/cryptography/json/Message.java
@@ -13,7 +13,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package org.javalaboratories.core.cryptography.transport;
+package org.javalaboratories.core.cryptography.json;
 
 import lombok.EqualsAndHashCode;
 import lombok.Value;
